refactor(domain): 패키지 변경
