+++ src/main/java/wooteco/subway/domain/fare/Fare.java
@@ -1,10 +1,11 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.fare;
 
 import java.util.Objects;
-import wooteco.subway.domain.strategy.discount.DiscountStrategy;
-import wooteco.subway.domain.strategy.discount.DiscountStrategyMapper;
-import wooteco.subway.domain.strategy.extrafare.ExtraFareStrategy;
-import wooteco.subway.domain.strategy.extrafare.ExtraFareStrategyMapper;
+import wooteco.subway.domain.Distance;
+import wooteco.subway.domain.fare.strategy.discount.DiscountStrategy;
+import wooteco.subway.domain.fare.strategy.discount.DiscountStrategyMapper;
+import wooteco.subway.domain.fare.strategy.extrafare.ExtraFareStrategy;
+import wooteco.subway.domain.fare.strategy.extrafare.ExtraFareStrategyMapper;
 
 public class Fare {
 
@@ -28,8 +29,8 @@ public class Fare {
 
     public Fare discountByAge(final int age) {
         final DiscountStrategy strategy = DiscountStrategyMapper.findStrategyBy(age);
-        final int discountAmount = strategy.calculate(value);
-        return new Fare(discountAmount);
+        final int discountedFare = strategy.calculate(value);
+        return new Fare(discountedFare);
     }
 
     public int getValue() {
