+++ src/main/java/wooteco/subway/service/PathService.java
@@ -6,10 +6,10 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Distance;
-import wooteco.subway.domain.Fare;
+import wooteco.subway.domain.fare.Fare;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Station;
-import wooteco.subway.domain.SubwayMap;
+import wooteco.subway.domain.subwaymap.SubwayMap;
 import wooteco.subway.dto.path.PathResponse;
 import wooteco.subway.exception.station.NoSuchStationException;
